ISSUE #2

+++ src/main/java/it/globus/finance/model/repo/TransactionRepo.java
@@ -3,7 +3,5 @@ package it.globus.finance.model.repo;
 import it.globus.finance.model.entity.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
-import java.math.BigInteger;
-
-public interface TransactionRepo extends CrudRepository<Transaction, BigInteger> {
+public interface TransactionRepo extends CrudRepository<Transaction, Long> {
 }
